+++ src/main/java/rbasamoyai/createbigcannons/datagen/CBCMixingRecipeProvider.java
@@ -2,13 +2,16 @@ package rbasamoyai.createbigcannons.datagen;
 
 import java.util.concurrent.CompletableFuture;
 
+import com.simibubi.create.AllItems;
 import com.simibubi.create.api.data.recipe.MixingRecipeGen;
 import com.simibubi.create.content.kinetics.mixer.MixingRecipe;
 import com.simibubi.create.content.processing.recipe.HeatCondition;
 import com.simibubi.create.content.processing.recipe.StandardProcessingRecipe;
+import com.simibubi.create.foundation.data.recipe.CommonMetal;
 
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
@@ -16,6 +19,7 @@ import net.minecraft.world.level.material.Fluids;
 import rbasamoyai.createbigcannons.CBCTags.CBCItemTags;
 import rbasamoyai.createbigcannons.CreateBigCannons;
 import rbasamoyai.createbigcannons.index.CBCItems;
+import rbasamoyai.createbigcannons.utils.CBCUtils;
 
 public class CBCMixingRecipeProvider extends MixingRecipeGen {
 
@@ -25,9 +29,34 @@ public class CBCMixingRecipeProvider extends MixingRecipeGen {
 
 	GeneratedRecipe
 
-		ALLOY_NETHERSTEEL_CAST_IRON = create(CreateBigCannons.resource("alloy_nethersteel_cast_iron"), b -> nethersteelAlloy(b, CBCItemTags.INGOT_CAST_IRON, 8, 8)),
+		ALLOY_NETHERSTEEL_CAST_IRON = create(CreateBigCannons.resource("alloy_nethersteel_cast_iron"), b -> nethersteelAlloy(b, CBCCommonMetal.CAST_IRON.ingots, 8, 8)),
 
-	ALLOY_NETHERSTEEL_STEEL = create(CreateBigCannons.resource("alloy_nethersteel_steel"), b -> nethersteelAlloy(b, CBCItemTags.INGOT_STEEL, 4, 8)),
+	ALLOY_NETHERSTEEL_STEEL = create(CreateBigCannons.resource("alloy_nethersteel_steel"), b -> nethersteelAlloy(b, CommonMetal.STEEL.ingots, 4, 8)),
+
+    // Adapted from Create: Alloyed.
+    ALLOY_BRONZE_TINLESS = create(CreateBigCannons.resource("alloy_bronze_tinless"), b -> b.require(CommonMetal.COPPER.ingots)
+        .require(CommonMetal.ZINC.ingots)
+        .require(AllItems.CINDER_FLOUR)
+        .requiresHeat(HeatCondition.HEATED)
+        .output(CBCItems.BRONZE_INGOT.get(), 2)),
+
+    ALLOY_BRONZE_TIN = create(CreateBigCannons.resource("alloy_bronze_tin"), b -> b.withCondition(CBCUtils.itemTagsPopulated(CommonMetal.TIN.ingots))
+        .require(CommonMetal.COPPER.ingots)
+        .require(CommonMetal.TIN.ingots)
+        .requiresHeat(HeatCondition.HEATED)
+        .output(CBCItems.BRONZE_INGOT.get(), 2)),
+
+    ALLOY_BRONZE_BRASS = create(CreateBigCannons.resource("alloy_bronze_brass"), b -> b.require(CommonMetal.BRASS.ingots)
+        .require(CommonMetal.BRASS.ingots)
+        .require(AllItems.CINDER_FLOUR)
+        .requiresHeat(HeatCondition.HEATED)
+        .output(CBCItems.BRONZE_INGOT.get(), 2)),
+
+    ALLOY_STEEL = create(CreateBigCannons.resource("alloy_steel"), b -> b.require(CommonMetal.IRON.ingots)
+        .require(CommonMetal.IRON.ingots)
+        .require(ItemTags.COALS)
+        .requiresHeat(HeatCondition.HEATED)
+        .output(CBCItems.STEEL_INGOT.get(), 2)),
 
 	CONGEALED_NITRO = create(CreateBigCannons.resource("congealed_nitro"), b -> b.require(Items.BLAZE_POWDER)
 		.require(Items.MAGMA_CREAM)