+++ src/main/java/rbasamoyai/createbigcannons/datagen/assets/CBCLangGen.java
@@ -21,6 +21,7 @@ import rbasamoyai.createbigcannons.CBCTags;
 import rbasamoyai.createbigcannons.CreateBigCannons;
 import rbasamoyai.createbigcannons.base.CBCRegistries;
 import rbasamoyai.createbigcannons.crafting.casting.CannonCastShape;
+import rbasamoyai.createbigcannons.datagen.CBCCommonMetal;
 import rbasamoyai.createbigcannons.index.CBCBlocks;
 import rbasamoyai.createbigcannons.index.CBCItems;
 import rbasamoyai.createbigcannons.utils.CBCUtils;
@@ -312,7 +313,6 @@ public class CBCLangGen {
 		.header("GAS MASK")
 		.summary("_Protects against gas clouds_ when worn on the head. _Does not protect_ against _lingering potions_, _dragon's breath_, or _potion fluid blobs_.");
 
-		// TODO: only run if on Fabric env
 		createEMITagTranslation(CBCTags.CBCItemTags.AUTOCANNON_AMMO_CONTAINERS);
 		createEMITagTranslation(CBCTags.CBCItemTags.AUTOCANNON_CARTRIDGES);
 		createEMITagTranslation(CBCTags.CBCItemTags.AUTOCANNON_ROUNDS);
@@ -320,10 +320,6 @@ public class CBCLangGen {
 		createEMITagTranslation(CBCTags.CBCItemTags.BIG_CANNON_PROJECTILES);
 		createEMITagTranslation(CBCTags.CBCItemTags.BIG_CANNON_PROPELLANT);
 		createEMITagTranslation(CBCTags.CBCItemTags.BIG_CANNON_PROPELLANT_BAGS);
-		createEMITagTranslation(CBCTags.CBCItemTags.BLOCK_BRONZE, "Bronze Blocks");
-		createEMITagTranslation(CBCTags.CBCItemTags.BLOCK_CAST_IRON, "Cast Iron Blocks");
-		createEMITagTranslation(CBCTags.CBCItemTags.BLOCK_NETHERSTEEL, "Nethersteel Blocks");
-		createEMITagTranslation(CBCTags.CBCItemTags.BLOCK_STEEL, "Steel Blocks");
 		createEMITagTranslation(CBCTags.CBCItemTags.CAN_BE_NITRATED, "Can be Nitrated");
 		createEMITagTranslation(CBCTags.CBCItemTags.DUST_GLOWSTONE, "Glowstone Dusts");
 		createEMITagTranslation(CBCTags.CBCItemTags.DUSTS_REDSTONE, "Redstone Dusts");
@@ -336,40 +332,21 @@ public class CBCLangGen {
 		createEMITagTranslation(CBCTags.CBCItemTags.HIGH_EXPLOSIVE_MATERIALS);
 		createEMITagTranslation(CBCTags.CBCItemTags.IMPACT_FUZE_HEAD, "Impact Fuze Head Components");
 		createEMITagTranslation(CBCTags.CBCItemTags.INEXPENSIVE_BIG_CARTRIDGE_SHEET, "Inexpensive Big Cartridge Sheets");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_BRASS, "Brass Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_BRONZE, "Bronze Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_CAST_IRON, "Cast Iron Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_COPPER, "Copper Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_IRON, "Iron Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_NETHERSTEEL, "Nethersteel Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.INGOT_STEEL, "Steel Ingots");
-		createEMITagTranslation(CBCTags.CBCItemTags.NITROPOWDER);
-		createEMITagTranslation(CBCTags.CBCItemTags.NITRO_ACIDIFIERS);
-		createEMITagTranslation(CBCTags.CBCItemTags.NUGGET_BRONZE, "Bronze Nuggets");
-		createEMITagTranslation(CBCTags.CBCItemTags.NUGGET_CAST_IRON, "Cast Iron Nuggets");
-		createEMITagTranslation(CBCTags.CBCItemTags.NUGGET_COPPER, "Copper Nuggets");
-		createEMITagTranslation(CBCTags.CBCItemTags.NUGGET_IRON, "Iron Nuggets");
-		createEMITagTranslation(CBCTags.CBCItemTags.NUGGET_NETHERSTEEL, "Nethersteel Nuggets");
-		createEMITagTranslation(CBCTags.CBCItemTags.NUGGET_STEEL, "Steel Nuggets");
-		createEMITagTranslation(CBCTags.CBCItemTags.SHEET_BRASS, "Brass Sheets");
-		createEMITagTranslation(CBCTags.CBCItemTags.SHEET_COPPER, "Copper Sheets");
-		createEMITagTranslation(CBCTags.CBCItemTags.SHEET_IRON, "Iron Sheets");
-		createEMITagTranslation(CBCTags.CBCItemTags.SHEET_STEEL, "Steel Sheets");
-		createEMITagTranslation(CBCTags.CBCItemTags.SPENT_AUTOCANNON_CASINGS);
-		createEMITagTranslation(CBCTags.CBCItemTags.STONE);
-		createEMITagTranslation(CBCTags.CBCItemTags.GAS_MASKS);
-		createEMITagTranslation(fabricTag("cast_iron_ingots"));
-		createEMITagTranslation(fabricTag("nethersteel_ingots"));
-		createEMITagTranslation(fabricTag("bronze_nuggets"));
-		createEMITagTranslation(fabricTag("cast_iron_nuggets"));
-		createEMITagTranslation(fabricTag("nethersteel_nuggets"));
-		createEMITagTranslation(fabricTag("steel_nuggets"));
-		createEMITagTranslation(fabricTag("ingots/cast_iron"), "Cast Iron Ingots (Forge Format)");
-		createEMITagTranslation(fabricTag("ingots/nethersteel"), "Nethersteel Ingots (Forge Format)");
-		createEMITagTranslation(fabricTag("nuggets/bronze"), "Bronze Nuggets (Forge Format)");
-		createEMITagTranslation(fabricTag("nuggets/cast_iron"), "Cast Iron Nuggets (Forge Format)");
-		createEMITagTranslation(fabricTag("nuggets/nethersteel"), "Nethersteel Nuggets (Forge Format)");
-		createEMITagTranslation(fabricTag("nuggets/steel"), "Steel Nuggets (Forge Format)");
+        createEMITagTranslation(CBCTags.CBCItemTags.NITROPOWDER);
+        createEMITagTranslation(CBCTags.CBCItemTags.NITRO_ACIDIFIERS);
+        createEMITagTranslation(CBCTags.CBCItemTags.SPENT_AUTOCANNON_CASINGS);
+        createEMITagTranslation(CBCTags.CBCItemTags.STONE);
+        createEMITagTranslation(CBCTags.CBCItemTags.GAS_MASKS);
+
+        createEMITagTranslation(CBCCommonMetal.BRONZE.storageBlocks.items(), "Bronze Blocks");
+        createEMITagTranslation(CBCCommonMetal.CAST_IRON.storageBlocks.items(), "Cast Iron Blocks");
+        createEMITagTranslation(CBCCommonMetal.NETHERSTEEL.storageBlocks.items(), "Nethersteel Blocks");
+		createEMITagTranslation(CBCCommonMetal.BRONZE.ingots, "Bronze Ingots");
+		createEMITagTranslation(CBCCommonMetal.CAST_IRON.ingots, "Cast Iron Ingots");
+		createEMITagTranslation(CBCCommonMetal.NETHERSTEEL.ingots, "Nethersteel Ingots");
+		createEMITagTranslation(CBCCommonMetal.BRONZE.nuggets, "Bronze Nuggets");
+		createEMITagTranslation(CBCCommonMetal.CAST_IRON.nuggets, "Cast Iron Nuggets");
+		createEMITagTranslation(CBCCommonMetal.NETHERSTEEL.nuggets, "Nethersteel Nuggets");
 
 		createEMICategoryTranslation("melting", "Basin Melting");
 		createEMICategoryTranslation("cannon_casting");